Added logging + other small changes
The program will write logs to "setting.log" if enabled in the setting
file (LOGGING_ENABLED 1). The log will only record when the program
starts, end, network loss, network recovery, and manual resets.

-Prevented potential CPU hog when the manual reset button is held down
for a long period of time

-Configuration file can now have comments, by beginning a line with '#'

-Values read in from the configuration are checked if they're within
valid ranges

-Added termination_routine, which is called when the program exists at
any point. For now, it simply prints out an exiting message and writes
it to the log.